Rename Макс и направленная ладья to  Max and directional boat | Макс и направленная ладья
En:
A rook is a chess piece that can move any number of cells horizontally and vertically.

Vladimir showed Max an unusual chessboard. It had a size of H × W (H contour lines and W verticals; horizontals are numbered from top to bottom from 1 to H, verticals are numbered from left to right from 1 to W), and the symbol U, L, D or R was written in each cell.

Vladimir explained to Max that this was a new puzzle he had invented. Vladimir put the rook in a cage at the intersection of the horizontal R1 and vertical C1. Now Max must bring this boat to the cell at the intersection of the horizontal R2 and vertical C2. But there is a limitation: the rook can still go to any number of cells, but only in the direction indicated on the cell in which it stood at the beginning of the turn:

U - the boat can only move up;
L - the rook can only move to the left;
D - the rook can only move down;
R - the rook can only move to the right.
Help Max get to the desired cell in the least number of moves.


Input format
The first line contains integers H and W (1 ≤ H, W ≤ 100) - the number of contours and the number of verticals of the board, respectively.

The following H lines describe the board. Each of them contains W characters Aij () - the directions indicated in the cells of the board.

The next line contains integers R1 and C1 (1 ≤ R1 ≤ H, 1 ≤ C1 ≤ W) - respectively the horizontal and vertical numbers, at the intersection of which the rook is originally located.

The next line contains integers R2 and C2 (1 ≤ R2 ≤ H, 1 ≤ C2 ≤ W) - respectively the horizontal and vertical numbers, at the intersection of which the rook should be.


Output format
Print a single integer - the minimum number of moves of the rook, after which it will be in the desired cell.

If the rook cannot reach the desired cell, print -1.



Ru:
Ладья — это шахматная фигура, которая может перемещаться на любое количество клеток по горизонтали и вертикали.

Владимир показал Максу необычную шахматную доску. Она имела размер H × W (H горизонталей и W вертикалей; горизонтали нумеруются сверху вниз от 1 до H, вертикали нумеруются слева направо от 1 до W), а в каждой клетке был написан символ U, L, D или R.

Владимир объяснил Максу, что это — придуманная им новая головоломка. Владимир поставил ладью в клетку на пересечении горизонтали R1 и вертикали C1. Теперь Макс должен довести эту ладью до клетки на пересечении горизонтали R2 и вертикали C2. Но есть ограничение: ладья всё ещё может ходить на любое количество клеток, но только в том направлении, которое указано на клетке, в которой она стояла в начале хода:

U — ладья может двигаться только вверх;
L — ладья может двигаться только влево;
D — ладья может двигаться только вниз;
R — ладья может двигаться только вправо.
Помогите Максу добраться до нужной клетки за наименьшее число ходов.

Формат ввода
Первая строка содержит целые числа H и W (1 ≤ H, W ≤ 100) — соответственно количество горизонталей и количество вертикалей доски.

Следующие H строк описывают доску. Каждая из них содержит W символов Aij () — направления, указанные в клетках доски.

Следующая строка содержит целые числа R1 и C1 (1 ≤ R1 ≤ H, 1 ≤ C1 ≤ W) — соответственно номера горизонтали и вертикали, на пересечении которых ладья находится изначально.

Следующая строка содержит целые числа R2 и C2 (1 ≤ R2 ≤ H, 1 ≤ C2 ≤ W) — соответственно номера горизонтали и вертикали, на пересечении которых ладья должна оказаться.

Формат вывода
Выведите одно целое число — минимальное количество ходов ладьи, после которых она окажется в нужной клетке.

Если ладья не сможет добраться до нужной клетки, выведите -1.